refactor: move mapping functions to dedicated modules
