+++ src/mapping/user.mapping.js
@@ -1,0 +1,17 @@
+const User = require('../models/user.model');
+
+function entityToModel(entity) {
+  if (!entity) return null;
+  return new User({
+    id: entity.id,
+    email: entity.email,
+    firstName: entity.first_name,
+    lastName: entity.last_name,
+    role: entity.role,
+  });
+}
+
+module.exports = {
+  entityToModel,
+};
+
