+++ src/services/regions.service.js
@@ -1,10 +1,5 @@
 const regionsRepository = require("../repositories/regions.repository");
-const Region = require("../models/region.model");
-
-function entityToModel(entity) {
-    if (!entity) return null;
-    return new Region({ id: entity.id, code: entity.code, name: entity.name });
-}
+const { entityToModel } = require("../mapping/region.mapping");
 
 exports.toModel = entityToModel;
